commitcast-backend/requirements.txt
# CommitCast Backend Dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.0
pydantic[email]>=2.0.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.0.0
spoon-ai-sdk
python-dotenv>=1.0.0
python-multipart>=0.0.6

